requirements.txt
requests
colorama
inquirer
python-dotenv
replicate
pillow
google-generativeai